requirements.txt
numpy==1.18.4
matplotlib==3.2.1
scipy==1.4.1
DendroPy==4.4.0
rpy2==3.3.3
biopython==1.76
psycopg2==2.8.5
scikit-learn==0.23.2
pytest-mock==3.3.1